Modify: main.js위치 변경
