+++ index.html
@@ -8,9 +8,7 @@
     <link rel="stylesheet" href="/src/styles/global.css" />
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.css" />
     <script src="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.js"></script>
-    <script type="module" src="/src/components/header/header.js"></script>
-    <script type="module" src="/src/components/footer/footer.js"></script>
-    <script type="module" src="/main.js"></script>
+    <script type="module" src="/src/pages/main/main.js"></script>
   </head>
 
   <body>
