--- src/styles/style.css
@@ -1,4 +1,0 @@
-#app {
-  width: 100%;
-  height: 100vh;
-}
